Updated SQL Scripts

+++ Bamazon-Schema.sql
@@ -10,7 +10,7 @@ CREATE TABLE products
 item_id INTEGER(11) AUTO_INCREMENT NOT NULL,
 product_name VARCHAR(100) NULL,
 department_name VARCHAR(100) NULL,
-price FLOAT(10) NULL,
+price DECIMAL(10,2) NULL,
 stock_quantity INTEGER(10) NULL,
 PRIMARY KEY (item_id)
 );
